progressive/stage2/streamlit-frontend/pyproject.toml
[project]
name = "streamlit-frontend"
version = "1.0.0"
description = "Streamlit frontend for dice roller with OpenTelemetry instrumentation"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.39.0",
    "requests>=2.32.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-requests>=0.48b0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
]
